I've been experimenting with vehicle on vehicle collision; how they push against eachother and even riding on top of the other. Still a work in progress (and I'll admit the code smells a little), but currently they can push against one another and even ride each other! There's still several things to work out, specifically dealing with getting pushed into tiles, but for the most part the difficult part is finished.

+++ RetroWar/RetroWar/Services/Implementations/AI/MoveLeftAndRightProcessor.cs
@@ -27,8 +27,8 @@ namespace RetroWar.Services.Implementations.AI
                 return false;
             }
 
-            enemy.OldX = (int)enemy.X;
-            enemy.OldY = (int)enemy.Y;
+            enemy.OldX = enemy.X;
+            enemy.OldY = enemy.Y;
 
             if (!enemy.CurrentActions.Contains(Models.Sprites.Actions.Action.Destroyed))
             {
